refactor(mcp): split feishu-context-mcp.ts into modular components (Issue #693)
- Split the 1259-line feishu-context-mcp.ts into smaller, focused modules
- Main file reduced from 1259 to 457 lines (~64% reduction)
- All 45 existing tests pass

## Changes

| File | Lines | Description |
|------|-------|-------------|
| `feishu-context-mcp.ts` | 457 | Entry point (was 1259) |
| `tools/types.ts` | 81 | Shared type definitions |
| `tools/send-message.ts` | 230 | send_user_feedback implementation |
| `tools/send-file.ts` | 181 | send_file_to_feishu implementation |
| `tools/card-interaction.ts` | 270 | update_card, wait_for_interaction |
| `tools/index.ts` | 21 | Module exports |
| `utils/card-validator.ts` | 72 | Feishu card validation |
| `utils/feishu-api.ts` | 117 | Low-level Feishu API |
| `utils/index.ts` | 8 | Utility exports |

## Benefits

- ✅ Better separation of concerns
- ✅ Easier testing of individual components
- ✅ Clearer responsibilities
- ✅ Improved maintainability
- ✅ Backward compatibility maintained (all exports preserved)

Fixes #693
