+++ src/mcp/tools/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Feishu MCP Tools - Module exports.
+ *
+ * This module re-exports all Feishu MCP tools for convenient importing.
+ */
+
+// Types
+export type {
+  MessageSentCallback,
+  SendUserFeedbackResult,
+  SendFileResult,
+  UpdateCardResult,
+  WaitForInteractionResult,
+  PendingInteraction,
+  FeishuCard,
+} from './types.js';
+
+// Tool implementations
+export { send_user_feedback, setMessageSentCallback, getMessageSentCallback } from './send-message.js';
+export { send_file_to_feishu } from './send-file.js';
+export { update_card, wait_for_interaction, resolvePendingInteraction } from './card-interaction.js';
