+++ src/mcp/utils/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Feishu MCP Utilities - Module exports.
+ *
+ * This module re-exports all Feishu MCP utility functions for convenient importing.
+ */
+
+export { isValidFeishuCard, getCardValidationError } from './card-validator.js';
+export { sendMessageToFeishu, createClient, notifyMessageSent } from './feishu-api.js';
